remove duplication in RubricTable
