+++ src/main/resources/frontend/src/views/PhaseView/RubricTable.vue
@@ -2,6 +2,7 @@
 import type {Rubric, TestResult} from "@/types/types";
 import ResultsPopup from "@/views/PhaseView/ResultsPopup.vue";
 import {ref} from "vue";
+import RubricTableRow from "@/views/PhaseView/RubricTableRow.vue";
 
 defineProps<{
   rubric: Rubric;
@@ -9,14 +10,6 @@ defineProps<{
 
 const selectedTestResults = ref<TestResult | undefined>(undefined);
 const selectedTextResults = ref<string | undefined>(undefined);
-
-const failedTests = (testResults: TestResult) => {
-  return testResults.numTestsFailed > 0 || testResults.numTestsPassed === 0;
-}
-
-const prettyQualityOutput = (qualityNotes: string) => {
-  return qualityNotes.replace(/✓/g, `<span class="success">✓</span>`).replace(/✗/g, `<span class="failure">✗</span>`);
-}
 
 </script>
 
@@ -30,51 +23,9 @@ const prettyQualityOutput = (qualityNotes: string) => {
         <th>Points</th>
         <th>Results</th>
       </tr>
-      <tr
-          v-if="rubric.passoffTests"
-          :class="failedTests(rubric.passoffTests.results.testResults) && 'failed'"
-      >
-        <td>{{rubric.passoffTests.category}}</td>
-        <td>{{rubric.passoffTests.criteria}}</td>
-        <td>{{rubric.passoffTests.results.notes}}</td>
-        <td>{{Math.round(rubric.passoffTests.results.score)}} / {{rubric.passoffTests.results.possiblePoints}}</td>
-        <td
-            class="selectable"
-            @click="() => {
-              selectedTestResults = rubric.passoffTests.results.testResults;
-              selectedTextResults = undefined;
-            }"
-        >Click here</td>
-      </tr>
-      <tr
-          v-if="rubric.unitTests"
-          :class="failedTests(rubric.unitTests.results.testResults) && 'failed'"
-      >
-        <td>{{rubric.unitTests.category}}</td>
-        <td>{{rubric.unitTests.criteria}}</td>
-        <td>{{rubric.unitTests.results.notes}}</td>
-        <td>{{Math.round(rubric.unitTests.results.score)}} / {{rubric.unitTests.results.possiblePoints}}</td>
-        <td
-            class="selectable"
-            @click="() => {
-              selectedTestResults = rubric.unitTests.results.testResults;
-              selectedTextResults = undefined;
-            }"
-        >Click here</td>
-      </tr>
-      <tr v-if="rubric.quality">
-        <td>{{ rubric.quality.category }}</td>
-        <td>{{ rubric.quality.criteria }}</td>
-        <td v-html="prettyQualityOutput(rubric.quality.results.notes)"></td>
-        <td>{{ Math.round(rubric.quality.results.score) }} / {{ rubric.quality.results.possiblePoints }}</td>
-        <td
-            class="selectable"
-            @click="() => {
-              selectedTestResults = undefined
-              selectedTextResults = rubric.quality.results.textResults;
-            }"
-        >Click here</td>
-      </tr>
+      <RubricTableRow :rubric-item="rubric.passoffTests" :results-clicked="() => {selectedTestResults = rubric.passoffTests.results.testResults;}"/>
+      <RubricTableRow :rubric-item="rubric.unitTests" :results-clicked="() => {selectedTestResults = rubric.unitTests.results.testResults;}"/>
+      <RubricTableRow :rubric-item="rubric.quality" :results-clicked="() => {selectedTextResults = rubric.quality.results.textResults;}"/>
     </table>
   </div>
 
@@ -96,7 +47,7 @@ const prettyQualityOutput = (qualityNotes: string) => {
   overflow-x: auto;
 }
 
-table, th, td {
+table, th {
   border: 2px solid #ddd;
 }
 
@@ -111,18 +62,4 @@ th {
   text-align: left;
 }
 
-td {
-  padding: 8px;
-  word-break: break-word;
-  white-space: pre-wrap;
-}
-
-.selectable:hover {
-  cursor: pointer;
-  background-color: #e1e1e1;
-}
-
-.failed {
-  background-color: rgba(255, 204, 204, 0.6);
-}
 </style>
